+++ joulescope/view.py
@@ -346,32 +346,10 @@ class View:
             stop = self._convert_time_to_samples(stop, units)
         return start, stop
 
-    def _samples_get(self, start=None, stop=None, units=None):
+    def _samples_get(self, start=None, stop=None, units=None, fields=None):
         s1, s2 = self._convert_time_range_to_samples(start, stop, units)
-        t1 = self.sample_id_to_time(s1)
-        t2 = self.sample_id_to_time(s2)
         self._log.debug('_samples_get(start=%r, stop=%r, units=%s) -> %s, %s', start, stop, units, s1, s2)
-        fields = ['current', 'voltage', 'power', 'current_range', 'current_lsb', 'voltage_lsb']
-        if self._stream_buffer.has_raw:
-            fields.append('raw')
-        sample_d = self._stream_buffer.samples_get(start=start, stop=stop, fields=fields)
-        result = {
-            'time': {
-                "range": {"value": [t1, t2], "units": "s"},
-                "delta": {"value": t2 - t2, "units": "s"},
-                'sample_id_range': {'value': [s1, s2], 'units': 'samples'},
-                "samples": {"value": s2 - s1, "units": "samples"},
-                'input_sampling_frequency': self._stream_buffer.input_sampling_frequency,
-                'output_sampling_frequency': self._stream_buffer.output_sampling_frequency,
-                'sampling_frequency': self._stream_buffer.output_sampling_frequency,
-            },
-            'signals': {},
-        }
-        for field, value in zip(fields, sample_d):
-            result['signals'][field] = {'value': value, 'units': FIELD_UNITS[field]}
-        if self._stream_buffer.has_raw:
-            result['signals']['raw']['voltage_range'] = self._stream_buffer.voltage_range
-        return result
+        return self._stream_buffer.samples_get(start=start, stop=stop, fields=fields)
 
     def _statistics_get(self, start=None, stop=None, units=None):
         """Get the statistics for the collected sample data over a time range.
@@ -430,7 +408,7 @@ class View:
     def stream_notify(self, stream_buffer):
         self._post('stream_notify', stream_buffer)
 
-    def samples_get(self, start=None, stop=None, units=None):
+    def samples_get(self, start=None, stop=None, units=None, fields=None):
         """Get exact samples over a range.
 
         :param start: The starting time.
@@ -438,8 +416,11 @@ class View:
         :param units: The units for start and stop.
             'seconds' or None is in floating point seconds relative to the view.
             'samples' is in stream buffer sample indicies.
+        :param fields: The fields to get.  None (default) gets the fundamental
+            fields available for this view instance, which may vary depending
+            upon the backend.
         """
-        args = {'start': start, 'stop': stop, 'units': units}
+        args = {'start': start, 'stop': stop, 'units': units, 'fields': fields}
         return self._post_block('samples_get', args)
 
     def statistics_get(self, start=None, stop=None, units=None, callback=None):